clean code; add sam2
